Another round of disambiguating int -> float or double promotion
svn path=/trunk/KDE/kdebase/workspace/; revision=747057

+++ kcontrol/keyboard/kcmmisc.cpp
@@ -212,7 +212,7 @@ void KeyboardConfig::delaySliderChanged (int value) {
 
 void KeyboardConfig::delaySpinboxChanged (int value) {
 	double alpha  = sliderMax / (log(5000.0L) - log(100.0L));
-	double logVal = alpha * (log(value)-log(100.0L));
+	double logVal = alpha * (log((double)value)-log(100.0L));
 
 	ui->delaySlider->setValue ((int)floor (0.5 + logVal));
